autofill/basic_form.html

<!DOCTYPE HTML>
<head>
<meta charset="utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/autofill_helpers.js"></script>
</head>
<body>
<form id="test_form">
  <input type="text" autocomplete="name" id="test_field">
</form>
<script>
  promise_test(async t => {
    if (!window.test_driver) {
      return;
    }
    const form = create_form_for_data({"name": "foobar"});
    await test_driver.save_to_auto_fill(form);
    // Now test that the values were properly filled
    const suggestions = await test_driver.get_autofill_suggestions(document.getElementById("test_form"), document.getElementById("test_field"));
    assert_equals(suggestions[0], "foobar");
  });
</script>
